Move stuff to explorer helpers
